Pull out CommonConversions class
